Hash module (#677)
* Create hash module

* Format

* Fix docs build

* Add docstrings

* Add pedersen_hash tests

* Format

* Fix test
