+++ starknet_py/hash/storage.py
@@ -1,8 +1,7 @@
 from functools import reduce
 
-from starknet_py.cairo.utils import _starknet_keccak
 from starknet_py.constants import ADDR_BOUND
-from starknet_py.utils.crypto.facade import pedersen_hash
+from starknet_py.hash.utils import _starknet_keccak, pedersen_hash
 
 
 def get_storage_var_address(var_name: str, *args: int) -> int:
