+++ starknet_py/hash/transaction_test.py
@@ -1,13 +1,13 @@
 import pytest
 
 from starknet_py.common import create_compiled_contract
-from starknet_py.tests.e2e.fixtures.misc import read_contract
-from starknet_py.utils.crypto.transaction_hash import (
+from starknet_py.hash.transaction import (
     TransactionHashPrefix,
     compute_declare_transaction_hash,
     compute_deploy_account_transaction_hash,
     compute_transaction_hash,
 )
+from starknet_py.tests.e2e.fixtures.misc import read_contract
 
 
 @pytest.mark.parametrize(
