+++ starknet_py/net/models/__init__.py
@@ -1,3 +1,3 @@
-from .address import Address, AddressRepresentation, compute_address, parse_address
+from .address import Address, AddressRepresentation, parse_address
 from .chains import StarknetChainId, chain_from_network
 from .transaction import Invoke, Transaction, compute_invoke_hash
