+++ starknet_py/net/udc_deployer/deployer.py
@@ -3,15 +3,16 @@ from __future__ import annotations
 import random
 from typing import List, NamedTuple, Optional, Union, cast
 
-from starknet_py.abi.parser import AbiParser
-from starknet_py.cairo.selector import get_selector_from_name
 from starknet_py.common import int_from_hex
+from starknet_py.abi.parser import AbiParser
 from starknet_py.constants import DEFAULT_DEPLOYER_ADDRESS, FIELD_PRIME
+from starknet_py.hash.address import compute_address
+from starknet_py.hash.selector import get_selector_from_name
+from starknet_py.hash.utils import pedersen_hash
 from starknet_py.net.client_models import Call, Hash
-from starknet_py.net.models import AddressRepresentation, compute_address, parse_address
 from starknet_py.serialization import serializer_for_function
+from starknet_py.net.models import AddressRepresentation, parse_address
 from starknet_py.utils.contructor_args_translator import translate_constructor_args
-from starknet_py.utils.crypto.facade import pedersen_hash
 from starknet_py.utils.sync import add_sync_methods
 
 ContractDeployment = NamedTuple(
