+++ starknet_py/tests/e2e/account/account_test.py
@@ -4,6 +4,7 @@ import pytest
 from starkware.starknet.public.abi import get_selector_from_name
 
 from starknet_py.contract import Contract
+from starknet_py.hash.address import compute_address
 from starknet_py.net.account.account import Account
 from starknet_py.net.account.base_account import BaseAccount
 from starknet_py.net.client_models import (
@@ -13,7 +14,7 @@ from starknet_py.net.client_models import (
     TransactionStatus,
 )
 from starknet_py.net.gateway_client import GatewayClient
-from starknet_py.net.models import StarknetChainId, compute_address
+from starknet_py.net.models import StarknetChainId
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
 from starknet_py.transaction_exceptions import TransactionRejectedError
