+++ starknet_py/tests/e2e/client/fixtures/prepare_network.py
@@ -7,7 +7,7 @@ from typing import AsyncGenerator, Dict, Tuple
 import pytest
 import pytest_asyncio
 
-from starknet_py.cairo.selector import get_selector_from_name
+from starknet_py.hash.selector import get_selector_from_name
 from starknet_py.net import AccountClient
 from starknet_py.tests.e2e.client.fixtures.prepare_net_for_gateway_test import (
     PreparedNetworkData,
