+++ starknet_py/tests/e2e/client/full_node_test.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from starknet_py.cairo.storage import get_storage_var_address
+from starknet_py.hash.storage import get_storage_var_address
 from starknet_py.net.client_errors import ClientError
