+++ starknet_py/tests/e2e/client/gateway_test.py
@@ -3,9 +3,9 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from starknet_py.cairo.selector import get_selector_from_name
-from starknet_py.cairo.storage import get_storage_var_address
 from starknet_py.common import create_compiled_contract
+from starknet_py.hash.selector import get_selector_from_name
+from starknet_py.hash.storage import get_storage_var_address
 from starknet_py.net.client_errors import ContractNotFoundError
 from starknet_py.net.client_models import (
     DeployTransaction,
