+++ starknet_py/tests/e2e/docs/account_creation/test_deploy_prefunded_account.py
@@ -11,10 +11,11 @@ async def test_deploy_prefunded_account(
 ):
     # pylint: disable=import-outside-toplevel, too-many-locals
     # docs: start
+    from starknet_py.hash.address import compute_address
     from starknet_py.net import KeyPair
     from starknet_py.net.account.account import Account
     from starknet_py.net.gateway_client import GatewayClient
-    from starknet_py.net.models import StarknetChainId, compute_address
+    from starknet_py.net.models import StarknetChainId
 
     # First, make sure to generate private key and salt
     # docs: end
