+++ starknet_py/utils/crypto/crypto_test.py
@@ -1,12 +1,8 @@
-import pytest
 from crypto_cpp_py.cpp_bindings import get_cpp_lib_file, unload_cpp_lib
 from starkware.crypto.signature.signature import private_to_stark_key, verify
 
-from starknet_py.utils.crypto.facade import (
-    compute_hash_on_elements,
-    message_signature,
-    use_cpp_variant,
-)
+from starknet_py.hash.utils import message_signature
+from starknet_py.utils.crypto.facade import use_cpp_variant
 
 
 def test_signing(monkeypatch):
@@ -36,20 +32,3 @@ def test_invalid_crypto_path(monkeypatch, mocker):
 
     file = get_cpp_lib_file()
     assert file is None
-
-
-@pytest.mark.parametrize(
-    "data, calculated_hash",
-    (
-        (
-            [1, 2, 3, 4, 5],
-            3442134774288875752012730520904650962184640568595562887119811371865001706826,
-        ),
-        (
-            [28, 15, 39, 74],
-            1457535610401978056129941705021139155249904351968558303142914517100335003071,
-        ),
-    ),
-)
-def test_compute_hash_on_elements(data, calculated_hash):
-    assert compute_hash_on_elements(data) == calculated_hash
