+++ starknet_py/utils/crypto/facade.py
@@ -1,10 +1,6 @@
-import functools
 import os
-from typing import Optional, Sequence
 
-from crypto_cpp_py.cpp_bindings import ECSignature, cpp_hash, get_cpp_lib_file
-from starkware.cairo.lang.vm.crypto import pedersen_hash as default_hash
-from starkware.crypto.signature.signature import sign
+from crypto_cpp_py.cpp_bindings import get_cpp_lib_file
 
 
 # Interface
@@ -14,25 +10,3 @@ def use_cpp_variant() -> bool:
     )
     cpp_lib_file = get_cpp_lib_file()
     return not force_disable_ext and bool(cpp_lib_file)
-
-
-def pedersen_hash(left: int, right: int) -> int:
-    if use_cpp_variant():
-        return cpp_hash(left, right)
-    return default_hash(left, right)
-
-
-def compute_hash_on_elements(data: Sequence):
-    """
-    Computes a hash chain over the data, in the following order:
-        h(h(h(h(0, data[0]), data[1]), ...), data[n-1]), n).
-
-    The hash is initialized with 0 and ends with the data length appended.
-    The length is appended in order to avoid collisions of the following kind:
-    H([x,y,z]) = h(h(x,y),z) = H([w, z]) where w = h(x,y).
-    """
-    return functools.reduce(pedersen_hash, [*data, len(data)], 0)
-
-
-def message_signature(msg_hash, priv_key, seed: Optional[int] = 32) -> ECSignature:
-    return sign(msg_hash, priv_key, seed)
